TGP-3381: eradicate warnings when running tests and it/tests, rename package goodsDescription
