+++ build.sbt
@@ -72,3 +72,9 @@ lazy val it =
 addCommandAlias("testAndCoverage", ";clean;coverage;test;it/test;coverageReport")
 addCommandAlias("prePR", ";scalafmt;test:scalafmt;testAndCoverage")
 addCommandAlias("preMerge", ";scalafmtCheckAll;testAndCoverage")
+inConfig(Compile)(scalacSettings)
+inConfig(Test)(scalacSettings)
+
+def scalacSettings = Seq(
+  scalacOptions := scalacOptions.value.distinct
+)
